Extract Obsidian vault folder-config reader out of theme
The Obsidian theme's .folder.yml reader used node:fs, which the worker
bundler can't satisfy on Cloudflare Workers. Move vault-walking into a
standalone CLI-only module and drop the corresponding hook from the
Theme interface so the worker bundle is free of node:fs imports.
