+++ packages/crawlers/package.json
@@ -7,7 +7,8 @@
   "types": "src/index.ts",
   "exports": {
     ".": "./src/index.ts",
-    "./themes": "./src/themes.ts"
+    "./themes": "./src/themes.ts",
+    "./obsidian/vault-folder-configs": "./src/obsidian/vault-folder-configs.ts"
   },
   "scripts": {
     "typecheck": "tsc --noEmit"
